// src/app/components/staff/types.ts
export type StaffObj = {
  staffid?: number;
  staffname?: string;
  contracttype?: string;
  contactno?: string;
  nic?: string;
  userid?: number;
  username?: string;
  role?: string;
  designation?: string;
  password?: string;
  country?: string;
};

export type UserType = {
  userid?: number;
  staffid?: number;
  username?: string;
};